+++ packages/file-entry-cache/tsconfig.json
@@ -2,8 +2,7 @@
   "compilerOptions": {
       "target": "ESNext",
       "module": "ESNext",
-      "moduleResolution": "node",                       /* Specify how TypeScript looks up a file from a given module specifier. */
-      "baseUrl": "./src",                                  /* Specify the base directory to resolve non-relative module names. */
+      "moduleResolution": "bundler",                    /* Specify how TypeScript looks up a file from a given module specifier. */
 
       /* Emit */
       "declaration": true,                              /* Generate .d.ts files from TypeScript and JavaScript files in your project. */
